server/requirements.txt
alembic==0.9.9
aniso8601==3.0.0
astroid==1.6.3
bcrypt==3.1.4
blinker==1.4
certifi==2018.1.18
chardet==3.0.4
click==6.7
coverage==4.5.1
coveralls==1.3.0
docopt==0.6.2
Flask==0.12.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.3
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
idna==2.6
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.0
marshmallow==2.15.0
mccabe==0.6.1
nose==1.3.7
psycopg2 ==2.7.4
psycopg2-binary==2.7.4
pycparser==2.18
PyJWT==1.6.1
pylint==1.8.4
python-dateutil==2.7.2
python-editor==1.0.3
pytz==2018.3
requests==2.18.4
six==1.11.0
SQLAlchemy==1.2.6
urllib3==1.22
webargs==2.1.0
Werkzeug==0.14.1
wrapt==1.10.11
